// src/api/generated/model/ambassadorControllerGetAmbassadorsParams.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Title
 * Description
 * OpenAPI spec version: 1.0
 */

export type AmbassadorControllerGetAmbassadorsParams = {
/**
 * Массив ID амбассадоров для фильтрации
 */
ambassadorIds?: string[];
/**
 * Массив ID комнат для фильтрации (амбассадор должен иметь связь AmbassadorRoom с accepted статусом)
 */
roomIds?: string[];
/**
 * Фильтр по имени (поиск в OrdPerson.name у связанной заявки). Работает только если больше 2 символов
 */
nameContains?: string;
/**
 * Фильтр по телефону (поиск в OrdPerson.phone). Работает только если больше 2 символов
 */
phoneContains?: string;
/**
 * Фильтр по ИНН (точное совпадение с OrdPerson.inn). Работает только если больше 2 символов
 */
innContains?: string;
/**
 * Номер страницы (от 1)
 */
page?: number;
/**
 * Размер страницы
 */
size?: number;
};